Manuals & fixing

+++ TaxOrg/Controllers/UploaderController.cs
@@ -20,7 +20,7 @@ using Organization = TaxOrg.Tools.Organization;
 
 namespace TaxOrg.Controllers
 {
-     public class UploaderController : Controller
+    public class UploaderController : Controller
     {
         //
         // GET: /MvcUploaderTest/Index
@@ -66,14 +66,13 @@ namespace TaxOrg.Controllers
 
                     try
                     {
-                        FileSystemRepository.FileDataName = "data";
-                        FileSystemRepository.FileTableName = "FsFile";
                         FileSystemRepository.FileSave(Request.Files[i], excelPath);
                     }
                     catch (Exception e)
                     {
                         e.SaveError();
-                        return new HttpStatusCodeResult(HttpStatusCode.InternalServerError, e.Message + ". " + st.error);
+                        return new HttpStatusCodeResult(HttpStatusCode.InternalServerError,
+                            e.GetErrorMessage() + ". " + st.error);
                     }
 
                 }
@@ -91,10 +90,10 @@ namespace TaxOrg.Controllers
                     TaxorgTools.CheckSaveTaxAccess();
 
                     using (var csvReader = new CsvReader<Organization>(excelPath, (e, row) =>
-                                                                                            {
-                                                                                                Buges.SaveBugRow(row, e.Message);
-                                                                                                e.SaveError();
-                                                                                            }))
+                    {
+                        Buges.SaveBugRow(row, e.Message);
+                        e.SaveError();
+                    }))
                     {
                         var taxRepository = new TaxRepository();
                         var firstOrg = csvReader.First();
@@ -102,7 +101,8 @@ namespace TaxOrg.Controllers
 
                         foreach (var organization in csvReader)
                         {
-                            taxRepository.SaveTaxToDb(organization.Inn, organization.TaxCode, organization.TaxName, organization.Date,
+                            taxRepository.SaveTaxToDb(organization.Inn, organization.TaxCode, organization.TaxName,
+                                organization.Date,
                                 organization.Tax);
                         }
 
@@ -119,7 +119,8 @@ namespace TaxOrg.Controllers
             //todo: add additional code to generate thumbnail for videos, associate files with entities etc
 
             //adding thumbnail url for jquery file upload javascript plugin
-            statuses.ForEach(x => x.thumbnailUrl = x.url + "?width=80&height=80"); // uses ImageResizer httpmodule to resize images from this url
+            statuses.ForEach(x => x.thumbnailUrl = x.url + "?width=80&height=80");
+            // uses ImageResizer httpmodule to resize images from this url
 
             //setting custom download url instead of direct url to file which is default
             statuses.ForEach(x => x.url = Url.Action("Index", "Org"));
@@ -132,7 +133,98 @@ namespace TaxOrg.Controllers
                 statuses.ForEach(x =>
                 {
                     //setting the error property removes the deleteUrl, thumbnailUrl and url property values
-                    x.error = rnd.Next(0, 2) > 0 ? "We do not have any entity with unlucky Id : '13'" : String.Format("Your file size is {0} bytes which is un-acceptable", x.size);
+                    x.error = rnd.Next(0, 2) > 0
+                        ? "We do not have any entity with unlucky Id : '13'"
+                        : String.Format("Your file size is {0} bytes which is un-acceptable", x.size);
+                    //delete file by using FullPath property
+                    if (System.IO.File.Exists(x.FullPath)) System.IO.File.Delete(x.FullPath);
+                });
+            }
+
+            var viewresult = Json(new {files = statuses});
+            //for IE8 which does not accept application/json
+            if (Request.Headers["Accept"] != null && !Request.Headers["Accept"].Contains("application/json"))
+                viewresult.ContentType = "text/javascript";
+
+            return viewresult;
+        }
+
+        public ActionResult UploadAllFiles(bool? inline, string ui = "bootstrap")
+        {
+            return View(inline);
+        }
+
+        [HttpPost]
+        public ActionResult UploadAllFiles(int? entityId)
+        {
+            // here we can send in some extra info to be included with the delete url 
+            var statuses = new List<ViewDataUploadFileResult>();
+            for (var i = 0; i < Request.Files.Count; i++)
+            {
+                string excelPath = null;
+                try
+                {
+                    var st = FileSaver.StoreFile(x =>
+                    {
+                        x.File = Request.Files[i];
+                        //note how we are adding an additional value to be posted with delete request
+                        //and giving it the same value posted with upload
+                        x.DeleteUrl = Url.Action("DeleteFile", new { entityId = entityId });
+                        x.StorageDirectory = Server.MapPath("~/Content/uploads");
+                        x.UrlPrefix = "/Content/uploads"; // this is used to generate the relative url of the file
+                        //                        x.UrlPrefix = ApplicationCustomizer.ExcelFilePath;
+                        // this is used to generate the relative url of the file
+                        //                        x.StorageDirectory = ApplicationCustomizer.ExcelFilePath;
+
+
+                        //overriding defaults
+                        x.FileName = Request.Files[i].FileName; // default is filename suffixed with filetimestamp
+                        x.ThrowExceptions = false;
+                        //default is false, if false exception message is set in error property
+                    });
+
+                    statuses.Add(st);
+                    excelPath = st.FullPath;
+
+                    try
+                    {
+                        st.name = FileSystemRepository.FileSave(Request.Files[i], excelPath).ToString();
+                    }
+                    catch (Exception e)
+                    {
+                        e.SaveError();
+                        return new HttpStatusCodeResult(HttpStatusCode.InternalServerError,
+                            e.GetErrorMessage() + ". " + st.error);
+                    }
+
+                }
+                catch (Exception e)
+                {
+                    e.SaveError();
+                }
+            }
+
+            //statuses contains all the uploaded files details (if error occurs then check error property is not null or empty)
+            //todo: add additional code to generate thumbnail for videos, associate files with entities etc
+
+            //adding thumbnail url for jquery file upload javascript plugin
+            statuses.ForEach(x => x.thumbnailUrl = x.url + "?width=80&height=80");
+            // uses ImageResizer httpmodule to resize images from this url
+
+            //setting custom download url instead of direct url to file which is default
+//            statuses.ForEach(x => x.name = );
+
+
+            //server side error generation, generate some random error if entity id is 13
+            if (entityId == 13)
+            {
+                var rnd = new Random();
+                statuses.ForEach(x =>
+                {
+                    //setting the error property removes the deleteUrl, thumbnailUrl and url property values
+                    x.error = rnd.Next(0, 2) > 0
+                        ? "We do not have any entity with unlucky Id : '13'"
+                        : String.Format("Your file size is {0} bytes which is un-acceptable", x.size);
                     //delete file by using FullPath property
                     if (System.IO.File.Exists(x.FullPath)) System.IO.File.Delete(x.FullPath);
                 });
@@ -146,7 +238,7 @@ namespace TaxOrg.Controllers
             return viewresult;
         }
 
-         //here i am receving the extra info injected
+        //here i am receving the extra info injected
         [HttpPost] // should accept only post
         public ActionResult DeleteFile(int? entityId, string fileUrl)
         {
@@ -155,7 +247,7 @@ namespace TaxOrg.Controllers
             if (System.IO.File.Exists(filePath))
                 System.IO.File.Delete(filePath);
 
-            var viewresult = Json(new { error = String.Empty });
+            var viewresult = Json(new {error = String.Empty});
             //for IE8 which does not accept application/json
             if (Request.Headers["Accept"] != null && !Request.Headers["Accept"].Contains("application/json"))
                 viewresult.ContentType = "text/plain";
@@ -175,5 +267,10 @@ namespace TaxOrg.Controllers
             }
         }
 
+        public ActionResult DownloadFile2(int idFile)
+        {
+            var file = FileSystemRepository.GetObjects().First(fs => fs.IdFileSystem == idFile);
+            return File(FileSystemRepository.GetStreamFromData(idFile), file.ContentType);
+        }
     }
 }
